requirements.txt
Flask
beautifulsoup4
requests
python-dateutil